Abins: Initial attempt to combine legacy 2D prototype with Mantid 4
Functionality to generate 2-d maps with Abins by selecting an
alternative "Instrument", building on work by Krzysztof Dymkowski

This is essentially a 3-way merge between the removed code from Mantid
3.9.2, a more recent prototype set of AbinsModules, and the current
Mantid code base. Some new development was needed to get Abins.py
working but otherwise this is left as intact as possible for testing,
benchmarking and further development.

Currently not to be used for production calculations. Succesfully
generates workspaces that can be viewed with the
MantidPlot spectrum viewer, but results are unverified.
Default precision of angular sweep has been reduced for speed.