Abins: dummy out 'threshold' logic which doesn't play well with 2D
I'm struggling to understand this logic. The docstrings and warnings
refer to "numerical stability" problems, but

      - the criterion is whether the maximum S in a quantum order is
        significantly in excess of a lower quantum order, assuming
        that increasing quantum orders will generally have lower
        intensity

      - the response is to exclude increasing numbers of small values
        from the spectrum. It's not easy to see how that will make a
        difference until the largest values are also excluded. Or why
        the outcome is a good thing.

This loop is sometimes triggered when running a 2D spectrum, added
huge amounts of runtime. It's not clear that the results of the
procedure are more physical than simply using a small fixed numerical
threshold.

Dummying out here in order to make progress 2D direct maps, but
inclined to remove the whole process as could affect the results and
the reasoning is not conveniently documented.