Abins CRYSTAL interface: make structure input more tolerant
Abins raises errors when loading from a recent molecular vibration
calculation from CRYSTAL17. This appears to be related to a small
difference in the format of the output block for atomic positions:

`     ATOM             X(ANGSTROM)         Y(ANGSTROM)         Z(ANGSTROM)`

compared to

`     ATOM          X(ANGSTROM)         Y(ANGSTROM)         Z(ANGSTROM)`

The relevant parsing function in GeneralAbInitioParser, "find_first"
is quite strict and uses the Python `in` keyword to match the
reference token to the output. I have added an alternative option to
this function to support the use of regular expressions, and modified
the CRYSTAL parser to use a regex which tolerates different amounts of
whitespace. The match is wrapped in '.*' because in some tests the data
line includes more whitespace and a '*' character.

The new regex setup should therefore be fully backwards compatible,
but does create some redundancy in the code. One option would be to
deprecate the 'msg' option that uses `in`, or to convert all 'msg'
inputs to a regex implementation automatically.