Abins: Add unit test, test data for CRYSTAL17 molecule
The process used to generate the existing reference data files for
Abins tests does not appear to be included in the Mantid repository.
An appropriate method is added to GeneralLoadAbInitioTester. As a pure
function/staticmethod this _could_ be located almost anywhere, but it
seems sensible to keep it near the method used for loading and
comparing the data.

Creating reference output with a given build of Mantid/Abins now takes
two lines of Python, e.g.::

    loader = AbinsModules.LoadCRYSTAL(ab_initio_filename='ref.out')
    AbinsModules.GeneralLoadAbInitioTester.save_ab_initio_test_data(
        loader, 'my_new_test')

which will write the 'my_new_test_data.txt' file and
'my_new_test_atomic_displacements_data_{k}.txt' files for each
phonon wavevector.