Abins broadening: fresh implementation alongside legacy
In order to work on the broadening in Abins, some adjustments are
being made to how these functions are called and it is made possible
to request different broadening schemes. A reimplementation of the
existing broadening is to be implemented as 'legacy' for ease of
comparison.

At this stage 'legacy' is the actual legacy implementation, while
'legacy_plus' is a similar implementation in the new framework. This
will replace the 'legacy' code as soon as possible.

The 'legacy' and 'legacy_plus' results give very similar results and
timing data, so while 'legacy_plus' does not provide full
compatibility it is a fair basis for comparison with new methods.