Abins: Some more broadening implementations
A considerable speedup is obtained by limiting the frequency range of the
Gaussian functions. Performance is comparable to the legacy
implementation - if a similar number of samples is used! 3*max(sigma)
corresponds to about 150 points, while the default number of samples
in the legacy implementation is just 50 points spread over the same
distance. If these numbers are brought into line, the performance is
quite similar. Interestingly, the for-loop summation and np.histogram
summation change their ranking for speed depending on the number of
samples. We should keep both approaches while developing this method,
and see which is fastest for production calculations.

The names aren't great and there's a lot of redundancy - this should
be cleaned up once the best approaches are established. There isn't
much reason to keep the legacy mode around either, as it gives
objectively wrong results.