Abins: remove GeneralData base class
Data objects in Abins inherit from a common base class "GeneralData".

This class does not implement any useful methods, it merely defines an
interface with `set()` and `extract()` methods. It also specifies that
derived classes should implement `__str__` -- but this is not
enforced (i.e. using Python's ABC features) and it is not clear that
the new placeholder (None) is superior to Python's default (which is,
at least, a string).

It is not clear that the standard interface is appropriate for all the
data classes:
- it seems appropriate that the AbinsData class (which represents the
  output of a pre-computed dynamics calculation) should be immutable,
  setting its content in __init__ with no set() method exposed.
- it seems appropriate that the AtomsData object make its data
  available directly through indexing, rather than require a method
  call every time data is extracted. Currently, other parts of Abins
  cache the output of extract() to streamline this.

There doesn't seem to be any part of Abins that makes use of this
consistent API for data objects; the HDF5 serialisation operates
on dictionaries rather than call extract() on objects. Perhaps a more
useful base class would replace extract() with a todict() method.

In order to make the above improvements to the API of particular
classes, the base class is removed in this commit. If IO refactoring
creates a use-case for a base class then it can be reimplemented at
that point; it might be that a metaclass or decorator is found to be
more suitable.