AbinsData: rework to be immutable
The API for AbinsData is changed so that the data is now set at
__init__ and no set() method is available for further mutation.

AbinsData represents the results of external calculations and so
should never need to change. By guaranteeing this we can perform
sanity checks once only, at initialisation, and simplify data access.

New tests are added for this object, using mocking to check data is
returned without unexpected intermediate processing.