Abins AtomsData: Rework to be immutable
Similar to work on AbinsData, the AtomsData can also be reworked to
remove set() and the need for further sanity-checking after
initialisation.

It's not clear why this object needs to store a nominal count of how
many atoms are present. This has been removed, along with some of the
related sanity checks. Inconsistencies will quickly become clear in
practice when combining this with vibrational data to construct an
AbinsData object. Really the file reader should take responsibility
for this question of "did I get all the data".

I have kept a sanity check for now that the dictionary keys form a
regular count up from zero, as this criterion is relied on
elsewhere. However, given that those 'atom_1' etc. are the _only_ keys
at the outer level of this data collection, it would seem more
efficient and elegant to represent it as a sequence and access the
items as integers.