WIP: Overhaul Abins KpointsData object
There are some curious design choices in the Abins KpointsData object;

- Low-frequency modes are removed from the data at the point that it
  is loaded
- This results in an uneven number of events at each k-point. As this
  is unsuitable for storage in a single array, the data is stored as
  dictionaries of arrays
- The dict entries are indexed by a string representation of an
  integer key

There seems to be another check on frequency vs the acoustic
threshold in SPowerSemiEmpiricalCalculator. Removing the pruning step
from KpointsData leads to inconsistent array sizes further down the
line.

It seems a more reasonable strategy to

- retain all data at the point of loading and store it in regular
  arrays
- Move frequency thresholding into the loop over k-points to avoid
  size-consistency problems

This is a cumbersome change that will break all the file-loading tests
as the internal data format is being changed.