Row-based implementation of Abins sums over element
Rework the function responsible for atom selection and summation by
isotope/element, from loop over a known nested data structure to a
"query"-like filter, sort and groupby of data rows.

This more flexible approach will support a wider range of data
divisions and metadata.

I think it's prettier, too, but that's a matter of opinion!