Database-like row access to Abins SData
Currently SData assumes that there will always be a heirarchy of
atom_index:quantum_order but this leaves no room for other properties
and will fall apart when we want to store coherent contributions that
are not attributed to a single atom.

A table-like structure should allow cleaner access to various
groupings of data. For now we just allow the table form to be
generated from existing structured data, but ultimately it could also
be used to initialise SData.

In this commit we use the table to implement creation of workspaces
from specified atom indices